[Paropt][Offload] Skip irrelevant constructs in device compilation.
For device compilation, the compiler must skip non-TARGET OMP constructs
that are not enclosed in TARGET regions and not in declare target functions.
Note that the compiler should not redundantly compile the non-TARGET construct
and later discard it, because that compilation may fail (besides incurring in
extra compile time unnecessarily). For example,
  #pragma omp for schedule(dynamic)
would trigger an assertion because dynamic scheduling is not supported for
device compilation.

Change-Id: Iad58ae5bee38175bebcda752d108b9ad3f9034c9
JIRA: CMPLRLLM-20558
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337179
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>