CMPLRLLVM-30549 and CMPLRLLVM-30514: Fix for flaky failure during dope vector const prop
This change set fixes a flaky failure when compiling
CPU2017/627.cam4_s. The main issue is that a possible
candidate for transpose can be passed to a vararg function
and it tries to collect the argument position. This
produces an assertion since the argument collected could be
larger than the max number of arguments.

These changes are similar to
https://git-amr-2.devtools.intel.com/gerrit/#/c/337195/
but for xmain-rel. The differences are that the patch
in https://git-amr-2.devtools.intel.com/gerrit/#/c/335598/
changes the analysis process but the issue can be replicable
and patch https://git-amr-2.devtools.intel.com/gerrit/#/c/336787/
introduced the debug print when there is no candidates for
transposing. This patch includes the print.

Change-Id: I42ab40440419983735a2740d334dac1ff7a3b6be
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337197
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>