CMPLRLLVM-30549 and CMPLRLLVM-30514: Fix for flaky failure during dope vector const prop
This change set fixes a flaky failure when compiling
CPU2017/627.cam4_s. The main issue is that a possible
candidate for transpose can be passed to a vararg function
and it tries to collect the argument position. This
produces an assertion since the argument collected could be
larger than the max number of arguments.

Change-Id: Ieb1b4cd8eedae62d8884785e53ecf2fd592db019
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337195
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>