Increase timeout of DirectoryWatcher unittest
This test was seen to be flaky in community testing, so they raised the
timeout from 3 seconds to 60 seconds. It's still flaky in our NetBatch
testing, so this commit raises the timeout even higher.

The test will normally complete in much less time, so this is unlikely
to increase total testing time.

This addresses CMPLRTST-13950

Change-Id: I1e0ef42dcd45e9e68621175cb9dc672086b790c2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337272
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>