[OptReport] Replace array with DenseMap for diagnostics table.
The diagnostics remark table which is used to map a unique
remark ID to its corresponding message is currently stored
as a static array, making it cumbersome to update with
out-of-order remark IDs. This patch replaces this array
data structure with a static DenseMap to address this
drawback.

JIRA: CMPLRLLVM-29333

Change-Id: I351912d70ed5a68c9555d9d9460f3f8c96ffb57f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337289
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>