[Paropt][Offload] Support interop-position in variant dispatch construct
In Fortran, when the variant function has string parameters, the compiler adds
special hidden parameters at the end of the function call for the string len.
In this case, we cannot append the interop_obj parameter at the end of the
function call. The FFE emits an interop position number in the declare variant
attribute, and the compiler emits the interop_obj in that position.

Change-Id: I02d08ec97400ceddaee1b35e6bbf40d7d86bcff9
JIRA: CMPLRLLM-29105
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337312
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>