CMPLRLLVM-24359: Distinguish between use of DVA for escape/constant propagation.
We need to distinguish between using dope vector analysis for escape analysis,
as we do for the transpose optimization and constant propagation.

In the case of escape analysis, we only need to trace the use of a dope vector
through all of its call sites to make sure there are no escapes. But in the
case of constant propagation, for each call, we need to ensure that there is
no other path reaching to that call which will pass down a different dope
vector array which may have different constant values.

Change-Id: Ia45915d30fe8666d0399758423b8aa3f00660ded
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337316
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>