[NFC] [CMPLRLLVM-29581] Make IP Cloning output IR deterministic
Make the IP Cloning pass insert the instructions in a deterministic
order so that the IR generated by the pass does not change from
run to run by making sets that are iterated upon to perform actions
rather than just used for queries use data structures that have a
deterministic iteration order. 'SplitInsts' is iterated upon to
insert new instructions, and 'Visited' is iterated up to populate
'SplitInsts'.

Change-Id: I4d347efe3c535a156dc93f38ef250147cdd0ecb6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337494
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>