[DTrans - OpaquePtr] DTransSafetyAnalyzer to set ComplexUse member of FieldInfo
The fields touched by a memory intrinsic function call that affects a
subset of the struture should be marked as ComplexUse. This is needed
to inform transformations that handle the "Memfunc partial write" safety
bit about fields accessed that do not have getelementptr instructions.
This makes the DTransSafetyAnalyzer behavior of the ComplexUse member
compatible with the behavior of the original DTransAnalysis pass.

Change-Id: I818019f6a72ab8519fc7868b3a7e91b0ede9b952
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337383
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>