[VPlan][TreeConflict] Introduce VPTreeConflict instruction to represent simple tree-conflict idioms.
Currently we analyze VConflict idioms in VPlan and lower them
directly into optimized Histogram if successfully recognized.
In this patch we introduce a new VPTreeConflict instruction
to represent all VConflict idioms which are either simple
Histograms or tree-conflicts. It supports scenarios where the
conflicting region has a single reduction-like operation.
This instruction will then be lowered into explicit VPlan IR
based on the type of idiom it represents. This is the first
patch in the series for the task to support simple tree-conflict
idiom in VPlan HIR vectorizer.

Change-Id: Ie62ed06ab50a5bea1a57f8e81d511e778cd093c6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337515
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>