[CMPLRLLVM-29700] Lower complex intrinsics (for scalar code only).
This lowering is not necessarily correct for all architectures, as the C ABI of
returning complex types can be complicated. In particular, the current
infrastructure is not entirely capable of handling the x86-32 ABI for returning
complex float.

It is thus necessary to limit the use of these intrinsics to architectures where
the lowering is verified (at this time, only x86-64).

Change-Id: Icc2564ee400f24988dd2d02eba112c2a1e18521b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334051
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Reviewed-by: Cranmer, Joshua <joshua.cranmer@intel.com>