[CMPLRLLVM-29703] Lower complex code to intrinsics in the Clang frontend.
For now, use of this lowering is hidden behind a cc1-only argument, which will
need to be passed on a per-architecture basis, as 32-bit complex intrinsics
are not fully correctly lowered (specifically, the i386 float _Complex ABI).

Additionally, the lowering is only enabled in cases that would require the call
to the __mul/div*c3 helper--i.e., complex-complex multiply and division by
complex. It may be prudent to lower the other cases as well, but as of this
patch, there are no optimizations for mixed complex/real arithmetic yet.

Change-Id: I66461b7471e0ebb49d005c38140e6002191252ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/334773
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>