[OpaquePtr] [NFCI] Add/populate BasePtrElementType field for GEP refs.
The patch introduces a BasePtrElementType field for refs which is populated
during parsing. For example, BasePtrElementType of the a ref like %A[i1][i2]
where %A is of type [10 x i32]* will be [10 x i32].

This patch should ideally be NFC but we have to give up parsing for certain
ptr phis when we cannot figure out the element type so it may not be NFC.

With opaque ptrs we cannot perform getPointerElementType() on the BaseCE
to get this type information. This information would be needed-

1) To correctly generate GEP for the ref during HIR code generation.
2) To access the ref's highest dimension type.

In subsequent changesets, we would need to add BasePtrElementType parameter to
DDRefUtils::createMemRef() and DDRefUtils::createAddressOfRef() which would
have to be provided by the callers.

Change-Id: I7fc7d511ebfab8efdf03e24619af1a7005b572e7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337717
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Yu, Lina <lina.yu@intel.com>
Reviewed-by: Choi, Yoonseo <yoonseo.choi@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>