CMPLRLLVM-29895: Fix inline report with metadata after argument promotion and dead arg elimination
This changeset introduces the method replaceCallBaseWithCallBase
for the metadata based inline report. This function will be used
to update the inline report information when a call is replaced
by another call.

Also, this change set fixes the issue that the metadata based
inline report was reporting incorrect information when argument
promotion and dead argument elimination was applied.

Change-Id: I130deb569b95fad534496d8c39e2d1f7280e2c20
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/337833
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>