[VPlan] Add support to create scalar live-in/out lists for HIR input.
This patch adds ScalarLiveInOutDescrHIR to track a DDRef that
represents a temp in input HIR which is live-in/out of the
target HLLoop. We also templatize VPLiveInOutCreator to work
for both llvm::Loop and loopopt::HLLoop and overload the
method that creates original live-in/out for each VPEntity.
This is the first patch in the series to enable CFGMerger
for HIR input.

Change-Id: Ia8e30a9ca1ab66bd67492d06aa4051342c43101d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338748
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>