[Paropt] FIRSTPRIVATE must imply TGT_MAP_PRIVATE.
Fixes Fortran cases for CMPLRLLVM-28092.  Fortran FE does not generate
MAP+FIRSTPRIVATE clauses for FIRSTPRIVATE structures, and Paropt
is not setting TGT_MAP_PRIVATE for them.  This prevents the by-val
optimization from kicking in, because it triggers on TGT_MAP_PRIVATE.

In general, this seems to be a bug in our Fortran toolchain: all FIRSTPRIVATE
items (not only aggregate) must use TGT_MAP_PRIVATE, otherwise, the value
passed to the target outlined routine may not be copied from the host
and taken from the current device copy.

Change-Id: I71138241fa5a6999c7d6cc804774c31a7f97e17e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338179
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>