[DTrans - OpaquePtr] Create a DynClone pass which supports opaque pointer IR
The DynClone pass implementation does not deal with pointer types directly,
but makes use the information collected by the DTrans analysis passes. This
requires updates to make use of the new DTrans infrastructure that supports
opaque pointers. This change reuses the basic dynamic cloning implementation
by providing it with a template class that provides the interface to the
DTrans information that is collected by either the DTransAnalysis pass or
the DTransSafetyAnalyzer pass by providing an adaptor of those results to
provide a common interface.

The bulk of this change is to port the test cases over to run with the new pass.
The tests cases are ported from the original dynamic cloning test cases and
extended to contain the metadata information used by the DTransSafetyAnalyzer,
and invoke the -dtrans-dyncloneop pass. Note, because the same implementation
code is being used, the debug trace is still named dtrans-dynclone.
Opaque pointer specific cases are added for cases that perform more detailed
of the transformed IR to check the handling when pointers are opaque, and
bitcasts are not present.

Change-Id: I6b3b3ad97dad0bc249c2abfa6785d30bf7696056
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338202
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>