[Driver][SYCL] Additional dependency gen step is part of offload toolchain (#4456)
The additional preprocessing step which is created that creates the
dependency information for the host compilation needs to be properly
setup as part of the SYCL offloading toolchain.  Without doing so, not
all of the dependencies (headers for instance) are properly pulled in.
Addresses:  CMPLRLLVM-30886
Cherry-pick of: https://github.com/intel/llvm/commit/6a4fa3a096bdfa679ac25dd7192d0762bfcf70e7

Change-Id: Idc7a6903f301d5e0fb4c84b95a602af6480a0326
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338438
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Artem Gindinson <artem.gindinson@intel.com>
Reviewed-by: Chilamantula, Harini <harini.chilamantula@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>