[Paropt] Replace lastprivate with private if item has no uses inside the region
This is the first part of the changes for simplifying lastprivate clauses which
changes shared privatization pass to replace lastprivate with private if item
has no uses inside the region. Next patches will include changes to analyse
lastprivate item uses outside the work region - if item is unused after the
region it should be safe to replace it with private.

Change-Id: I17e3f1ea2d782fdc4b863e04c25923b52a714748
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338647
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>