[SYCL] Fix SYCL config processing for DPC++ RT (#4366)
Fix SYCL configuration file processing for DPCPP runtime.
The reason for this change is that the config for SYCL broke
from the simplest changes, for example, from a comment in
the config. Without these changes, it was a dangerous code
that may fail and even don't send an error.
What's new:

 1. Rewritten line processing;
 2. Supported format:
 3. ConfigName=Value #comment;
 4. Skip lines which do not contain variables;
 5. Throw error message if variable is incorrect;
 6. Several tests for new functionality.

Change-Id: I78b07bfac63417e98fb5cec4425d90f796ce8eb2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338722
Reviewed-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>
Tested-by: Kabanov Denis <denis.kabanov@intel.com>
Reviewed-by: Mikhail Lychkov <mikhail.lychkov@intel.com>