[VPlan][HIR] Fix incorrect assumption in loop upper bound decomposition
VPlan HIR decomposer had an incorrect assumption that if the result
of loop upper bound decomposition is a VPInstruction then it is newly
generated. The attempt to blindly set the underlying node causes an
assertion fail as a result. The fix here removes the assumption by
setting the underlying node only when new instructions are generated
during loop upper bound decomposition.

Change-Id: Iecec8219bd24d892d19f899a3253512da05e654d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338946
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>