Bugfix to add missing pointer dimension in the contracted alloca ref.
This ref is formed during HIRCrossLoopArrayContraction pass.
The issue was caught while working on opaque ptrs changes in parsing.

The issue is currently hidden because CG has logic to add missing
zeroes in GEPs but that logic doesn't work with the opaque ptr changes.

It is not clear whether the CG logic needs to exist but we should
probably add stricter checks to catch such issues.

Change-Id: I5f51ad298a63056f0ef2fdcc2e0b900487110f5f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338950
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Yu, Lina <lina.yu@intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>