[VPlan][HIR] Restrict stride improvements using underlying IR to inner loops
Stride information improvement using underlying HIR is done using the loop
level at which the memory reference occurs. This, however, needs to be done
using the level of the loop where we are vectorizing at. While this can be
done, we restrict such improvements to inner loops where we saw a
performance need before. In general, DA should be improved so that we can
avoid looking at underlying IR. The outer loop support can be added when
we see a real need.

Change-Id: If989753ead96edf448fcc59c1ac375cb4c7c7199
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/338966
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>