Ensure field-annotations on pointers properly match the AS of the field
Discovered in SYCL, the field annotations were always cast to an i8*,
which is an invalid bitcast for a pointer type with an address space.
This patch makes sure that we create an intrinsic that takes a pointer
to the correct address-space and properly do our casts.

It's a cherry-pick of https://reviews.llvm.org/D109003
Merged conflicts are resolved in intel/llvm:
commit 6abb6b1d46aa9705360aabf4be56716d1973e994
Merge: 8f3cc7b34c54 42ae7eb581ff
Author: Erich Keane <erich.keane@intel.com>
Date:   Wed Sep 1 09:29:34 2021 -0700

    Merge from 'main' to 'sycl-web' (#6)

      CONFLICT (content): Merge conflict in
clang/lib/CodeGen/CodeGenFunction.cpp

Change-Id: I771866f070f5a9eb36183c24d94e42ebf8ec8a38
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/339145
Reviewed-by: Rao, Premanand M <premanand.m.rao@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>