[VPlan][HIR][NFCI] Restructure Label/Goto creation in HIR vector code generation (#1)
When creating gotos for block terminators during HIR vector code generation,
creation of the target labels for the gotos was deferred until we saw the
target block during code generation. We now go ahead and create the target
label right away if one does not exist. This change is being done in
preparation for formation of proper HLIFs.