Implement flag -sox and Qsox (#72)
This is a partial implmentation of -sox functionality. This
implementation saves compilation options in executable file.
Clang saves compiler version in executable file by default.

Not implemented - Keywords which lets users choose whether to
include lists of inlined and -prof-use compiled functions. At this
time it is unclear if/when this will be implemented.

Epic tracking FE/Driver/BE work -
https://jira.devtools.intel.com/browse/CMPLRLLVM-24924