[VPlan] Add support to create multiple VPlans in CFGMerger for HIR input. (#20)
* [VPlan] Add support to create multiple VPlans in CFGMerger for HIR input.

This patch lays the foundation in CFGMerger to create additional
VPlans for HIR input based on chosen VecScenario. We introduce
VPPeelRemainderHIR instruction to represent scalar HLLoops, it has
an internal temp initialization map which specifies what VPValues
must be used to initialize a temp in preheader of scalar loop. We
also templatize and create new fabs in CFGMerger to support HIR
input. VPOrigLiveOut is also templatized and extended to support
different live-out types (llvm::Value or loopopt::DDRef). This is
the third patch in the series to enable CFGMerger for HIR input.

Change-Id: Ifa60883e702749f1e62556a37a183081a7abb330

* Rename templated typenames for VPOrigLiveOut and Loop.

This commit updates the original changes to use existing VPlan
naming scheme - using "HIR" suffix for HIR path and no suffix
for LLVM-IR path. It also resolves typename used for different
loop types used as parameter to templated functions.