[VPlan][HIR] Initial outer loop vectorization support in VPlan HIR path (#65)
* [VPlan][HIR] Initial outer loop vectorization support in VPlan HIR path

The changes in this patch add the initial support for outer loop vectorization
in the VPlan HIR path. The changes are off by default. We currently generate
an unknown loop as soon as we encounter a new inner VPLoop. Optimizing these
unknown loops to DO loops when possible will be implemented later. We also
do not recognizer inner loop IVs currently. As a result, canon expressions
which use such IVs will not be linear currently. This is also planned for
later.

* Add and use naive version of if/else generation when CFG merger is enabled

* Use new command line switch to control naive if/else generation