[CMPLRLLVM-31233] [DTrans] Fix AllocAnalyzer to avoid entering an infinite loop (#105)
The allocation analyzer offset analysis code would enter an infinite loop if
any binary operator other than an 'Add' instruction was encountered because
it would fail to update the loop control variable in that case. This change
fixes that problem, and adds braces to the code to conform to the coding
convention that braced bodies should be all or none.