[Paropt] Enable inlining for all functions called from target regions (#14)
This patch adds new module pass vpo-paropt-target-inline which
recursively adds alwaysinline attribute to all functions called
from target regions. This pass is added to pipeline when internal
option -enable-vpo-paropt-target-inline is specified (default is
false).

This is intended to be used as a workaround for CMPLRLLVM-30502
until the problem is fixed.