[CMPLRLLVM-30856] Enable array contraction for target-cpu=alderlake (#81)
This changeset implements the functional (code logic) part of the ticket CMPLRLLVM-30856 (which is a perf-regression ticket). Target-CPU alderlake was not used (not available) during development, and no alderlake perf-testing is performed. Instead, -xalderlake (ADL_1c) and -xskylake (SKL_1c) flags were used on a cascade lake machine to verify the correctness of the changeset. Original (before) as well as expected (after) behavior, in terms of HIR output, is described in the following GitHub comment:

https://github.com/intel-restricted/applications.compilers.llvm-project/pull/81#issuecomment-922152007