[VPlan][HIR][NFCI] Fix excessive compile time issue for cpu2017/521 (#47)
A recent change in VPExternalDef construction for HIR started using HIR dump methods
to setup the name of these external defs. Unfortunately, dumping HIR can cause dump
methods for blobs(SCEV expressions) to be invoked which has the side effect of causing
AsmWriter slot assignment mechanism to kick in. The slot numbers assigned are used
in printing LLVM IR Values without a name. The slot assignment can be very expensive
in terms of compile time especially when we are dealing with an lto compilation of
huge applications such as 521.wrf. Ideally, the slot assignment penalty should be
paid only when we are going to dump IR. This was done so that SOA analysis could
report meaningfuld names in its analysis dumps.

To fix the excessive compile time problem, this patch implements changes to look at
LLVM Value of temp Blobs(SelfBlob) to check if they have a name and to use that name
as the name of VPExternalDef. These are the names of interest which were being used
to improve SOA analysis dumps.

JIRA :: CMPLRLLVM-31086