[Driver] Prevent IR generation with -save-temps for product builds (#267)
The product/release builds (INTEL_PRODUCT_RELEASE) do not allow for
usage of -emit-llvm-bc.  When using -save-temps, each intermediate
step during the compilation is enabled (preprocessing->llvm-ir->
assembly->object).  For product release builds, allow for the llvm-ir
step to be collapsed so that we perform preprocessing->assembly->object
and do not emit an error.