[OpaquePtr][HIRTempCleanup] Removed call to getPointerElementType() and added lit testing. (#257)
Previously, HIRTempCleanup was ignoring load candidates for substitution which have
function pointer types.With opaque pointers we will not know from the load whether we are
loading a function pointer. The main use site where we wanted to disallow substitution was
function pointer of indirect call. This is now being checked explicitly when analyzing uses.
We allow substitution in other valid insts like stores. I added a new lit test to verify this
behavior.