[NFC][VPlan] Refactor VPSubscriptInst (#217)
None of the existing code uses it for LLVM IR path to represent
llvm.intel.subscript intrinsic and it *is* focused on being able to represent
loopopt::RegDDRef's information. Don't try to pretend otherwise in interfaces.
Before this patch sizeof(VPSubscriptInst) == 304
After it sizeof(VPSubscriptInst) == 336

That looked strange to me because I expected the new approach to storing data to
require less size so I continued digging. A hint was here (state before this
patch):
  sizeof(VPSubscriptInst::DimStructOffsetsMapTy) == 32
  sizeof(VPSubscriptInst::DimTypeMapTy) == 32

That means that we had zero pre-allocated StructOffsets storage and even the
first offset would result in a significant heap allocation! I've verified it by
running opt on hir_vplan_decomp_memop_struct_offsets_1.ll under gdb. Indeed,
      // Get type associated for dimension.
      Types[I - 1] = Ref->getDimensionType(I);
in VPDecomposerHIR::decomposeMemoryOp causes the DenseMap to grow resulting in a
call to
  llvm::allocate_buffer (Size=2560, Alignment=8)
when the first StructOffset data is added to the map.
The new approach also eliminates map lookups changing them to trivial array
indexing.