[VPlan] Fix LIT test issues (#413)
Recent xmain-web pulldown caused cost modeling changes where we now vectorize
a loop that was not being vectorized before. This exposed an issue where the
incoming loop in the LIT test was not in LCSSA form. To avoid CM dependencies,
the test now forces vectorization with VF=4 and the loop has been updated to
be in LCSSA form.