Parsing and Sema support for OpenMP 5.1 ompx_places clause (#259)
This is a proposed standardized version of Intel's existing 'subdevice'
clause.

ompx_places( [domain-modifier,] start [ : length [ : stride ] ] )

where:

'domain-modifier' can be one of
  numa_domain
  subnuma_domain
If absent, default value is numa_domain.

'start' is a nonnegative integer expression and must be present.
'length' is a positive integer expression. If absent, default value is 1.
'stride' is a positive integer expression. If absent, default value is 1.

The IR (eventual) IR has all four operands.

The 'domain-modifier' is represented as a constant unsigned integer:
  0 for numa_domain
  1 for subnuma_domain

Partially implements: CMPLRLLVM-29293