CMPLRLLVM-30334: Types merging for C and C++ with mangled names (Part 1) (#409)
This change set introduces the concept of merging the types using
the mangled names. The basic idea is that the IR mover will collect
a type from the source module module, check the mangled name,
will try to find the same type that uses the same mangled name
in the destination module and will merge them. The result module
(lto.o) will be a module without repeated types that are presented
with alternate names (e.g. %struct.TESTSTRUCT and %struct.TESTSTRUCT.0),
with the exception of certain cases.

Also, this change set fixes the issue where the CFE expresses a
field that is a function pointer with an empty structure. The IR
mover will use the DTrans metadata to reconstruct the type during
the merging process in order to move the type with the actual
result.

NOTE: This patch is a fix for PR #324. It add's the Demangle libraries
for the build process.