[OpaquePtr][NFC] BitCastDestTy field now stores pointer element type rather than a pointer type. (#350)
BitCastDestTy which used to store a pointer type will not contain element type
information when we move to opaque pointers. The element type is needed to
compute the size of load/store refs among other things.

To resolve this, we now store the pointer element type instead.
One exception to the rule is vector AddressOf refs which store a vector of
pointers. The field has been renamed to BitCastDestVecOrElemTy.