Codegen for new OpenMP clause: ompx_places (#455)
Emit IR for proposed standardized version of Intel's existing 'subdevice'
clause. The IR is mostly equivalent to 'subdevice' IR except for difference
in first argument. For 'subdevice', it is a non-negative integer constant.
For 'ompx_places', it is one of two domain-modifiers: numa_domain or
subnuma_domain.

Partially Implements: CMPLRLLVM-29293