[CMPLRLLVM-30698] Handle profile mapping when lines are hoisted out of loops (#441)
The community code avoids reusing source line information when a line is hoisted
out of a loop in order to maintain the ability to map profile information
collected with autoFDO to source positions because otherwise the count within
the loop looks the same as the count outside the loop when looking at counts
based on source line. xmain has turned off the behavior to track line positions
in the opt-report (CMPLRLLVM-7678). This change allows the line information
to be tracked for the opt-report, but also uses the debug discriminator field
that is used for distinguishing between source line numbers that occur in
different basic blocks to allow the profiling information to be mapped.
The code is based on the code within MIRAddFSDiscriminators to identify
discriminators in use to generate the discriminator value to use when hoisting
the instruction.