[SYCL] Refactor C++14 support (#406)
* [SYCL] Refactor C++14 support (#4303)

1. Modified some constructs to avoid using C++17 features
2. Fixed the test to better check support of different C++ versions
3. Fixed several compiler warnings
4. Add a warning(message) if headers are compiled in <= C++17 mode

* [SYCL] Fix is_device_copyable with range rounding (#4478)

Suppose we have the following code:
```
template <> struct is_device_copyable<DeviceCopyable>
                   : std::true_type {};
...
  DeviceCopyable DevCop(0);
  Q.submit([=](sycl::handler& cgh){
    const sycl::range<2> range(1026, 1026);
    cgh.parallel_for(range,[=](sycl::item<2> item) {
      (void)DevCop;
    });
  });
```
This code doesn't compile because range rounding optimization
wraps kernel lambda function, so we have something like this:
```
|- WrapperLambda
 |-KernelLambda
 ||-DevCop
 |-Other Wrapper captures
```
According to the implementation of is_device_copyable,
we check whether all the fields of WrapperLambda are device
copyable. KernelLambda is not device copyable since there is
no corresponding template specialization of is_device_copyable.

It's not possible to provide a template specialization for
is_device_copyable with kernel lambda since lambda types don't
have the name. To fix this issue, we create a functor class
RoundedRangeKernel and provide an is_device_copyable trait for this
class, which simply forwards check to nested kernel lambda.

Co-authored-by: DenisBakhvalov <61807338+DenisBakhvalov@users.noreply.github.com>