[VPlan][HIR] Use getWidenedType instead of FixedVectorType::get (#344) (#451)
This patch uses getWidenedType instead of FixedVectorType::get as the latter requires
a Scalar type to be passed in. We were hitting an assert when trying to create the
RegDDRef for a deconstructed phi copy instruction as we had an implicit assumption
that the value being copied has a scalar type. The fix replaces all uses of
FixedVectorType::get with use of getWidenedType which accounts for vector type
being passed as an input argument. The changes fix first part of the issue reported
in CMPLRLLVM-31488.

Jira :: CMPLRLLVM-31488