[Dtrans] Fix DTransMD for windows padding (#481)
CMPLRLLVM-31701 pointed out an example where we don't properly emit the
metadata for padding (and assert!) around a bitfield.  Windows bitfields
seem to have padding possible both before and after with little in the
way of guidance as to how it works.

Additionally, there is some tail-padding that exists sometimes, so fix
that for windows as well.