[VecClone] Create llvm.directive.region.entry manually (#487)
The abstraction in IntrinsicUtils isn't that helpful and having full control
would simplify later change to emit "Typed" version of the OMP clauses in
parallel with existing spelling. Also, those "Typed" version don't even have
utilities to create them yet.

This patch changes two things:
1) The order in which clauses are emitted - no grouping anymore, instead they're
presented in the same order as vector attribute's parameters' description.
2) Emit separate clause for each uniform value. First, that is still legal.
Second, it's the only choice for the "Typed" version of the clause.
Not strictly an NFC because now I split the preheader unconditionally, even if
there is no linears/uniforms as that allows to simplify the code even further.
It is going to be changed and/or simplified by VPOCFGRestructuring/CFGSimplify
anyway.