[OpenCL] Fix unable to find declaration of OpenCL builtins (#569)
Community introduced a helper function `getOpenCLCompatibleVersion()` to
simplify conditions on OpenCL C and C++ for OpenCL versions. As community
has not fully supported OpenCL C 3.0, the conditions for OpenCL versions
to judge if a builtin is supported only has two case: equals to 200 or
not. But we have a patched clang that supports OpenCL C 3.0, so we should
take this case into consideration.