[CMPLRLLVM-31867]: Bugfix to distinguish scalar AddressOf ref with vector bitcast from vector AddressOf ref. (#644)
BitCastDestVecOrElemTy can either contain-
1) Element type of the bitcast, Or
2) A vector of pointers for a vector AddressOf ref. This is only
generated by vectorizer because we don't currently support vector GEPs.

In cases where we had a bitcast to a pointer to vector like <4 x i32>*
in the incoming IR, we were incorrectly treating it as case 2) when it
should be treated as case 1). To fix this we now check whether any
of the dimension indices of the ref is a vector type.