CMPLRLLVM-31891: Remove DTrans tests dependent on -std-link-opts option (#701)
The community code has removed the option '-std-link-opts' of 'opt' which
ran the legacy LTO pass manager, https://reviews.llvm.org/D109664.
These tests were using that option to verify that DTrans analysis and
transformation passes get run when needed (2 of them had the option removed
by a pulldown coordinator in an attempt to fix them). These tests were not
providing much value from checking that DTrans passes are run in a particular
order. Also, the checks that pass dependencies get run is really a test of
the pass managers and not DTrans specific. Therefore, these tests are being
removed, rather than trying to re-introduce the code into xmain that the
community removed.